Added PASS_QUAD implementation. (possibly WIP, not tested) Several threading-related files weren't being included by the CMake script Updated some doxygen documentation. Removed Light::_getNearClipVolume & _getFrustumClipVolumes as they weren't being used (they were needed by shadow volumes) Optimized Math::saturate & lerp

+++ OgreMain/include/OgrePass.h
@@ -563,7 +563,7 @@ namespace Ogre {
 		void addTextureUnitState(TextureUnitState* state);
         /** Retrieves a pointer to a texture unit state so it may be modified.
         */
-        TextureUnitState* getTextureUnitState(unsigned short index);
+        TextureUnitState* getTextureUnitState( size_t index );
         /** Retrieves the Texture Unit State matching name.
             Returns 0 if name match is not found.
         */
